fix: change show code modal state management through redux and use react useState
