+++ packages/pn-personafisica-webapp/src/redux/delegation/reducers.ts
@@ -10,8 +10,6 @@ import {
   revokeDelegation,
   openAcceptModal,
   closeAcceptModal,
-  openVerificationCodeModal,
-  closeVerificationCodeModal,
 } from './actions';
 import { Delegation } from './types';
 
@@ -40,7 +38,7 @@ const delegationsSlice = createSlice({
     verificationCodeModalState: {
       open: false,
       name: '',
-      verificationCode: ''
+      verificationCode: '',
     },
   },
   reducers: {},
@@ -76,16 +74,6 @@ const delegationsSlice = createSlice({
     builder.addCase(closeRevocationModal, (state) => {
       state.modalState.id = '';
       state.modalState.open = false;
-    });
-    builder.addCase(openVerificationCodeModal, (state, action)=> {
-      state.verificationCodeModalState.verificationCode = action.payload.verificationCode;
-      state.verificationCodeModalState.open = true;
-      state.verificationCodeModalState.name = action.payload.name;
-    });
-    builder.addCase(closeVerificationCodeModal, (state) => {
-      state.verificationCodeModalState.open = false;
-      state.verificationCodeModalState.name = '';
-      state.verificationCodeModalState.verificationCode = '';
     });
     builder.addCase(revokeDelegation.fulfilled, (state, action) => {
       state.modalState.open = false;
